api/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
google-generativeai
chromadb
beautifulsoup4
requests
python-dotenv
